+++ tests/test_producers.py
@@ -156,6 +156,18 @@ class MethodTestCase(TestCase):
         self.assertEqual(result, "hello, tester!")
 
 
+class RelatedFieldTestCase(TestCase):
+    def test_related_field(self):
+        owner = Owner.objects.create(name="test owner")
+        Widget.objects.create(name="test 1", owner=owner)
+        Widget.objects.create(name="test 2", owner=owner)
+        Widget.objects.create(name="test 3", owner=owner)
+
+        produce = producers.related_field("widget_set", "name")
+        result = produce(owner)
+        self.assertEqual(result, ["test 1", "test 2", "test 3"])
+
+
 class PKListTestCase(TestCase):
     def test_pk_list(self):
         owner = Owner.objects.create(name="test owner")