+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- New `producers.related_field` and `pairs.related_field` functions for fetching the value of any field from a related object or objects.
+
 ## [1.0.0] - 2020-10-13
 
 Initial stable release.